+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased (2.0.0) (XXXX-XX-XX)
 ---------------------
 
+- Add getFeatureInfo via backend proxy for wms layers.
+
 - Add region aggregation as a fourth aggregation tool. Draw regions and get
   raster aggregations when clicked.